Rename Polish programmer's layout to Polish layout
None of the Linux, MacOS, Windows triple implements Polish programmers's layout
as described in the Polish Standard PN-I-06000:1997.

What's worse is that deviations from the standard are not consistent across
the aforementioned systems. Probably the safest common denominator are Polish
diacritic letters.

Because this layout differs from US ANSI just by these letters now and because
it looks like US-ANSI-derived layout is the most often used one in Poland,
it's better to call it Polish layout.

Signed-off-by: Piotr Orzechowski <piotr@orzechowski.tech>